requirements.txt
aniso8601==7.0.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SeaSurf==0.2.2
Flask-WTF==0.14.2
google-api-core==1.14.2
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pydantic==1.8.2
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
WTForms==2.2.1
